refactor(api): move update health reports into machine controller
Relocate host and DPU firmware update health-report helpers from the machine
update manager module into the machine state controller.

These helpers are used by machine state handling rather than the update manager
itself, so colocating them with the machine controller clarifies ownership and
reduces coupling as preparation for extracting the machine state controller.
