+++ crates/api/src/tests/dpu_nic_firmware.rs
@@ -26,6 +26,7 @@ use model::machine_update_module::{
 use crate::CarbideResult;
 use crate::machine_update_manager::dpu_nic_firmware::DpuNicFirmwareUpdate;
 use crate::machine_update_manager::machine_update_module::MachineUpdateModule;
+use crate::state_controller::machine::health_report::create_host_update_health_report_dpufw;
 use crate::tests::common;
 use crate::tests::common::api_fixtures::TestManagedHost;
 use crate::tests::common::api_fixtures::test_managed_host::TestManagedHostSnapshots;
@@ -286,7 +287,7 @@ async fn test_clear_completed_updates(
         .await
         .unwrap();
 
-    let health_override = crate::machine_update_manager::machine_update_module::create_host_update_health_report_dpufw();
+    let health_override = create_host_update_health_report_dpufw();
     // Mark the Host as in update.
     db::machine::insert_health_report(
         &mut txn,
