+++ crates/api/src/tests/power_shelf_state_controller/error_state.rs
@@ -29,11 +29,9 @@ use db::power_shelf as db_power_shelf;
 use forge_secrets::credentials::TestCredentialManager;
 use model::power_shelf::{PowerShelf, PowerShelfControllerState, PowerShelfMaintenanceOperation};
 use sqlx::PgConnection;
+use state_controller::db_write_batch::DbWriteBatch;
+use state_controller::state_handler::{StateHandler, StateHandlerContext, StateHandlerOutcome};
 
-use crate::state_controller::db_write_batch::DbWriteBatch;
-use crate::state_controller::state_handler::{
-    StateHandler, StateHandlerContext, StateHandlerOutcome,
-};
 use crate::tests::common::api_fixtures::create_test_env;
 use crate::tests::common::api_fixtures::site_explorer::new_power_shelf;
 use crate::tests::power_shelf_state_controller::fixtures::power_shelf::{
